Final Updates for Abood
