+++ app/contact/page.js
@@ -24,8 +24,10 @@ export default function ContactPage() {
           <div className="text-center">
             <h1 className="text-4xl md:text-5xl lg:text-6xl font-bold mb-6">Contact Us</h1>
             <p className="text-xl md:text-2xl mb-8 text-gray-100 max-w-3xl mx-auto">
-              Have a question or want to get in touch? Fill out the form below or use our contact
-              details to reach us directly.
+              <strong>Have a question or need assistance?</strong><br />
+              If you’re a dental practice looking for skilled professionals, we’re here to help.<br />
+              <strong>Are you a dental professional looking for work?</strong><br />
+              We’d love to connect with you. Use the form below or our contact details to get in touch.
             </p>
             <p className="text-lg md:text-xl mb-8 text-gray-100 max-w-2xl mx-auto">
               <strong>Opening Hours:</strong><br />
